move replay_buffer to memory folder
